Added district class and triangle splitting almost works
